#54: Fixing Docker ports issue :whale:
